refactor(dashboard/connectors): extract connector-state cluster from the 2446-line surface
connector-status.ts was a 2446-line route surface. Move the pure
ConnectorStateLabel cluster — the label type/guard and its visual mappings
(connectorStateLabel, connectorCardBorderClass, connectorStateTone,
dotClassForLabel, connectorCardStateClass, connectorStatusPill{Class,Label}) —
into connector-state.ts. These are pure functions of a GateConnectorInfo /
ConnectorStateLabel with no component/signal state.

connector-status.ts re-exports the externally-imported members
(connectorStateLabel, connectorCardBorderClass, ConnectorStateLabel) for
back-compat, matching the ./connector-constants precedent, so no importer
changes. Surface drops from 2446 to 2348 lines; the extracted cluster (120
lines) is unit-tested in isolation by the existing connector-card-border test.

Behavior unchanged: connector card-border / status / overview-strip tests (120)
pass without modification. tsc clean for both files, eslint clean.
